compile: Move dummyFS into separate file
